requirements-dev.txt
ipython
django-extensions
-r requirements.txt